Move constructPatchOpValueForMultiValuedAttribute method to utils
